refactor: remove ava
