+++ server/package.json
@@ -11,26 +11,16 @@
   "engines": {
     "node": ">=12.0.0"
   },
-  "ava": {
-    "extensions": [
-      "ts"
-    ],
-    "require": [
-      "ts-node/register"
-    ]
-  },
   "main": "main.js",
   "scripts": {
-    "test": "ava --timeout=121s",
-    "test-new": "jest",
+    "test": "jest",
     "test:watch": "jest --watch",
     "start": "npm run dev",
     "build": "tsc -p .",
     "dev": " SPACES_DEFAULT_BUCKET_NAME=dev.2anki.net SPACES_ENDPOINT=fra1.digitaloceanspaces.com WORKSPACE_BASE=/tmp/w nodemon --watch '**/*.ts' --exec \"ts-node\" server.ts",
     "lint": "tslint  -e \"node_modules\" -c tslint.json '**/*.ts' '**/*.tsx'",
     "lint:fix": "tslint --fix -e \"node_modules\" -c tslint.json '**/*.ts' '**/*.tsx'",
-    "purge-js": "rm `find . -name '*.js'|grep -v node_modules`",
-    "tt": "ava --watch --timeout=121s"
+    "purge-js": "rm `find . -name '*.js'|grep -v node_modules`"
   },
   "repository": {
     "type": "git",
@@ -85,7 +75,6 @@
     "@types/node": "^14.14.36",
     "@types/node-fetch": "^2.5.10",
     "@types/react-dom": "17.0.11",
-    "ava": "^3.15.0",
     "dotenv": "^10.0.0",
     "jest": "^27.5.1",
     "nodemon": "^2.0.7",
